requirements.txt
discord.py == 1.3.3
python-dotenv == 0.13.0
regex == 2017.4.5
async == 0.6.2
